program will now err out automatically if input file does not have .nlp in it

+++ make_text_output_directory.py
@@ -14,25 +14,28 @@ __version__='output_text1.0'
 import os,shutil,sys
 
 def main(input_file):
-    output_directory=input_file[:input_file.find('.nlp')]
-    if(os.path.isdir(output_directory)):        
-        try:
-            os.rmdir(output_directory)
-            os.mkdir(output_directory)
-            return  (dict,{'errorType':'Warning','errorString':'Output directory already existed at program runtime. It was empty and was deleted'})
-        except:
+    if '.nlp' in input_file:
+        output_directory=input_file[:input_file.find('.nlp')]
+        if(os.path.isdir(output_directory)):        
             try:
-                shutil.rmtree(output_directory)
+                os.rmdir(output_directory)
                 os.mkdir(output_directory)
-                return (dict,{'errorType':'Warning','errorString':'Output directory already existed at program runtime. It was not empty and was deleted'})
+                return  (dict,{'errorType':'Warning','errorString':'Output directory already existed at program runtime. It was empty and was deleted'})
             except:
-                return(Exception,'FATAL ERROR: Output directory already existed at program runtime: '+output_directory+'. It could not be deleted'+str(sys.exc_info()[1]))
-            
-    else:
-        try:        
-            os.mkdir(output_directory)
-            return None, None
-        except:            
-            return(Exception,'FATAL ERROR: failed to create directory: '+output_directory+' based on input filename '+input_file)
+                try:
+                    shutil.rmtree(output_directory)
+                    os.mkdir(output_directory)
+                    return (dict,{'errorType':'Warning','errorString':'Output directory already existed at program runtime. It was not empty and was deleted'})
+                except:
+                    return(Exception,'FATAL ERROR: Output directory already existed at program runtime: '+output_directory+'. It could not be deleted'+str(sys.exc_info()[1]))
+                
+        else:
+            try:        
+                os.mkdir(output_directory)
+                return None, None
+            except:            
+                return(Exception,'FATAL ERROR: failed to create directory: '+output_directory+' based on input filename '+input_file)
 
-    
+    else:
+        return(Exception,'FATAL ERROR: bad input file name: '+input_file+' program aborted')
+
